debian: don't rely on an environment variable in /etc/default to start
Using an environment variable to check if we should start the daemon is
now considered bad practice. If the user installs the daemon, they want
it to start with a "good" default configuration. Since it is quite
difficult to have a good default configuration with ExaBGP, we just
don't ship a default /etc/exabgp/exabgp.conf and won't start if the file
is not present. As soon as the user adds its own configuration file, the
daemon will be started.

If a user doesn't want to start the daemon, they should use the
appropriate mechanisms for that. For example, `systemctl disable
exabgp.service`.

exabgp.env is now generated with the default settings from ExaBGP and we
ask the user for confirmation before erasing (through ucf). There is no
more reading configuration from /etc/default/exabgp. This simplifies the
install/removal scripts.

systemd service file is also updated to start only if the configuration
file exists.