healthcheck: refactor a bit to decrease complexity
pylint complains about too many branches and statements. Unfortunately,
there is little we can do about that. The number of statements in the
`loop()` function is high because it should be a module of its own but
we want to use a single-contained script which is incompatible with a
module. Internal functions are used to looks like a module but pylint
doesn't care.

The too many branches is here because this is how a finite-state
automaton should be implemented in Python. The automaton is too simple
to warrant the use of distinct functions for each state.