Don't enforce logging to stdout with systemd
If a user configures ExaBGP to log to syslog or to a file, this won't
work when using systemd. Moreover, the default value for log.destination
is already stdout. At least, logging to syslog would also work as
systemd will provide a per-service /dev/log. If the user decides to log
to a file, it will lose the ability to see logs through systemd.