healthcheck: don't use enum module from Python3.4+
The API is different from "our" implementation. The official API expects
a name and a list of names (or a string of space-separated names, or
something else). ALways use "our" implementation.

This was done on master in c3a137e231963de8c5ca63bc0325b0344292b5e4 but
the commit cannot be easily cherry-picked.

Close #367.