Some low-hanging items / non-code impacting
All changes were done in comments / non-code with the exception of a few strings displayed.

1. Where possible, update HTTP URLs to HTTPS to encourage more secure browsing. Most of them were auto-rewrite/redirects these days, so this saves a little bit of browser I/O.

All of the URLs were tested manually. If there was not an HTTPS version or it had a private certificate it was not adjusted to HTTPS.

Marked a URL as (historic) to indicate that the URL is not longer alive, but may be accessible in various historic searches.
(note that the cisco URL in README.md doesn't work but I didnt adjust it since README.md is the public face)

2. Minor spelling fixes:
  recevied -> received  (lib/exabgp/reactor/protocol.py)
  Indendifier -> Identifier (lib/exabgp/bgp/message/open/capability/negotiated.py)

3. Sprinkled a few URLs for RFC or ID references for easier click-n-code