Fix busy spin when receving commands on named pipe
When opening the read named pipe in RD_ONLY, the select loop will always
return that something is available to read from self.r_pipe once
something has opened and closed the fifo like `exabgpcli` or `echo "foo" >
/path/to/fifo`. Unfortunately, those are going to be EOF and exabgp will
be spinning trying to read content from the fifo when nothing is
actually available.
To work around that, we can open the reader fifo in RDWR mode, this will
ensure that the pipe will always have a potential writer and no EOF will
be available for select.

Fixes #704