Remove Debian packaging from master branch
Provide instructions on how to build using the debian/sid branch. The
drawback of such an approach is that it may not work on older versions
of Debian/Ubuntu. If this happens, it would be possible to maintain a
derivative branch for such versions. This will still be easier to
maintain.