fix parsing of labelled default IP route
The current check <24 is off by one and should be <=24,
which breaks parsing in the case of a prefix of lenth zero
(default route).

This change fixes this test.

This change also makes rd_mask a constant because the size
of this field does not depend on whether we are VPNv4 or VPNv6;
RD is always 8 bytes.