Only complain on RIB injection if no peer can take it #1179
The behaviour was to return an issue if any of the peers could
not take the RIB. With the new `neighbor *` feature bulk sending
routes should now be accepted.

If any peer can accept the route as the family was negotiated,
then the announcement was a success, only a failure if no peers
could take it.