improve counters/statistics
Reworks the statistics code in a backward compatible way. A new options
called statistics is added and set to true by default.

It shows at every keep-alive sent what are the latest counters for
packet types and others.

The commit is longer than should as some ruff formating got in.